+++ candidateSelection.h
@@ -1,6 +1,6 @@
 #include "lists.h"
 
-double costFunction(struct candidate cand1, struct candidate cand2);
-double* candidateSelection(struct candidateList **windowList, long length);
-void candidateSelectionSegment(double* fundamentals, 
+float costFunction(struct candidate cand1, struct candidate cand2);
+float* candidateSelection(struct candidateList **windowList, long length);
+void candidateSelectionSegment(float* fundamentals, 
 				struct candidateList **windowList, long final, long start);
