+++ findCandidates.c
@@ -4,15 +4,15 @@
 #include "findCandidates.h"
 
 // https://stackoverflow.com/questions/6514651/declare-large-global-array
-static double ratioRanges[15] = {1.15, 1.29, 1.42, 1.59,  
-				  1.8,  1.9,  2.1,  2.4,
-				  2.6,  2.8,  3.2,  3.8,
-				  4.2,  4.8,  5.2};
+static float ratioRanges[15] = {1.15, 1.29, 1.42, 1.59,  
+				1.8,  1.9,  2.1,  2.4,
+				2.6,  2.8,  3.2,  3.8,
+				4.2,  4.8,  5.2};
 
-static double mRanges[15] = { 4,  3,  2,  3,  
-			     -1,  1, -1,  2,
-			     -1,  1, -1,  1,  
-			     -1,  1, -1};
+static float mRanges[15] = { 4,  3,  2,  3,  
+			    -1,  1, -1,  2,
+			    -1,  1, -1,  1,  
+			    -1,  1, -1};
 
 
 struct orderedList calcCandidates(double* peaks, int numPeaks)
@@ -24,7 +24,7 @@ struct orderedList calcCandidates(double* peaks, int numPeaks)
 
 	for (i=0; i<numPeaks-1;i++){
 		for (j=i+1; j<numPeaks; j++){
-			m = calcM(peaks[i],peaks[j]);
+			m = calcM((float)peaks[i],(float)peaks[j]);
 			if (m>0) {
 				orderedListInsert(&candidates,peaks[i]/m);
 			}				
@@ -33,10 +33,10 @@ struct orderedList calcCandidates(double* peaks, int numPeaks)
 	return candidates;
 }
  
-double calcM(double f_i, double f_j){
+double calcM(float f_i, float f_j){
 	// find the index, i, of the right most value in ratioRanges less than
         // f_j/f_i. it returns mRanges[i]
-	double ratio = f_j/f_i;
+	float ratio = f_j/f_i;
 
 	// basically using algorithm from python's bisect left and then moving
 	// one index to the left. This could probably be improved
@@ -54,8 +54,8 @@ double calcM(double f_i, double f_j){
 
 
 struct candidateList* distinctCandidates(struct orderedList* candidates,
-					 int max_length, double xi,
-					 double f0Min, double f0Max)
+					 int max_length, float xi,
+					 float f0Min, float f0Max)
 {	
 	// Finds the distinct candidates from the list of candidates. This is
         // done by selecting the candidate with the most other candidates
@@ -102,7 +102,8 @@ struct candidateList* distinctCandidates(struct orderedList* candidates,
 		// add that candidate to distinct
 		if ((candidates->array[maxIndex] >= f0Min) &&
 		    (candidates->array[maxIndex] <= f0Max)){
-			candidateListAdd(distinct, candidates->array[maxIndex],
+			candidateListAdd(distinct,
+					 (float)(candidates->array[maxIndex]),
 					 confidence[maxIndex]);
 		}
 		// determine first: index of first entry in candidates that is
