+++ findCandidates.h
@@ -1,5 +1,5 @@
 struct orderedList calcCandidates(double* peaks, int numPeaks);
-double calcM(double f_i, double f_j);
+double calcM(float f_i, float f_j);
 struct candidateList* distinctCandidates(struct orderedList* candidates,
-					 int max_length, double xi,
-					 double f0Min, double f0Max);
+					 int max_length, float xi,
+					 float f0Min, float f0Max);
